refactor: Merged Serializer And Deserializer
