+++ tests/temp.py
@@ -1,1 +1,1 @@
-print(bool(1231234))
+print("1" <= "00" <= "26")
